#1326 Phase 1: extract worker_loop body to worker/loop_body/mod.rs
The ~1278-LOC `worker_loop` function body is moved verbatim from
userspace-dp/src/afxdp/worker/mod.rs (lines 995-2273) to a new
worker/loop_body/mod.rs. The original site adds
`mod loop_body; pub(crate) use loop_body::worker_loop;` and drops
the body.

Refs #1326 (Phase 1).

Result:
- worker/mod.rs LOC: 2635 → 1359 (under the 2000-LOC modularity
  gate that triggered #1326).
- Directory module layout established: worker/loop_body/.
- Pure code motion. No behavioral changes. No signature changes.
- pub(crate) fn worker_loop is re-exported from worker/mod.rs so
  every existing call site (worker_runtime, coordinator) keeps
  working unchanged.

Test plan:
- cargo check clean with no new warnings.
- cargo test --release: 1487 tests pass.
  (One pre-existing failure in snat_contract_documents_current_fail_closed_runtime
  is unrelated — same failure occurs on origin/master commit 936b076d,
  triggered by a stale doc-guard expecting "fail-closed" wording in
  docs/userspace-dataplane-gaps.md that was never added on master.)

Plan + adversarial review:

Plan doc: docs/pr/1326-worker-loop-extract/plan.md (v4.0 — final
across 6 review rounds).

- Codex r6 (task-mpmwdrx4-kpdreh): PLAN-READY on commit 6f384430.
- AGY r4 (review-mpmw4xjc-92bdff): PLAN-READY on commit 7a56fb71
  (still valid on v3.4 — v3.3→v3.4 was wording-only).

Per-stage carve-out into setup.rs + tick.rs + poll_drive.rs +
debug_report.rs (the eventual 5-file target architecture documented
in plan.md) is deferred to follow-up PRs. Reasons in plan §"Deferred
to follow-up tickets": risk isolation (file-level move is provably
semantic-equivalent; per-stage carves are real refactors), reviewer
concerns scoped (LLVM borrow shape, hybrid inlining etc. only apply
when sub-fns exist), and incremental landing matches #959 / #1189
multi-phase pattern.
